break stream helper into pieces
